+++ odbc-api/src/handles/error.rs
@@ -15,8 +15,13 @@ pub enum Error {
     /// SQL Error had been returned by a low level ODBC function call. A Diagnostic record is
     /// obtained and associated with this error.
     Diagnostics(DiagnosticRecord),
+    /// A user dialog to complete the connection string has been aborted.
     #[error("The dialog shown to provide or complete the connection string has been aborted.")]
     AbortedConnectionStringCompletion,
+    #[error("ODBC diver manager does not seem to support the required ODBC version 3.80. (Most
+        likely you need to update unixODBC if you run on a Linux. Diagnostic record returned by
+        SQLSetEnvAttr:\n{0}")]
+    OdbcApiVersionUnsupported(DiagnosticRecord),
 }
 
 pub trait IntoResult {